+++ Integrations/AWS-EC2/AWS-EC2.py
@@ -2636,6 +2636,117 @@ def modify_image_attribute_command(args):
         demisto.results('Image attribute sucessfully modified')
 
 
+def detach_internet_gateway_command(args):
+    client = aws_session(
+        region=args.get('region'),
+        roleArn=args.get('roleArn'),
+        roleSessionName=args.get('roleSessionName'),
+        roleSessionDuration=args.get('roleSessionDuration'),
+    )
+    kwargs = {}
+    if args.get('InternetGatewayId') is not None:
+        kwargs.update({'InternetGatewayId': args.get('InternetGatewayId')})
+    if args.get('VpcId') is not None:
+        kwargs.update({'VpcId': args.get('VpcId')})
+
+    response = client.detach_internet_gateway(**kwargs)
+    if response['ResponseMetadata']['HTTPStatusCode'] == 200:
+        demisto.results('Subnet sucessfully deleted')
+
+
+def delete_subnet_command(args):
+    client = aws_session(
+        region=args.get('region'),
+        roleArn=args.get('roleArn'),
+        roleSessionName=args.get('roleSessionName'),
+        roleSessionDuration=args.get('roleSessionDuration'),
+    )
+    kwargs = {}
+    if args.get('SubnetId') is not None:
+        kwargs.update({'SubnetId': args.get('SubnetId')})
+
+    response = client.delete_subnet(**kwargs)
+    if response['ResponseMetadata']['HTTPStatusCode'] == 200:
+        demisto.results('Subnet sucessfully deleted')
+
+
+def delete_vpc_command(args):
+    client = aws_session(
+        region=args.get('region'),
+        roleArn=args.get('roleArn'),
+        roleSessionName=args.get('roleSessionName'),
+        roleSessionDuration=args.get('roleSessionDuration'),
+    )
+    kwargs = {}
+    if args.get('VpcId') is not None:
+        kwargs.update({'VpcId': args.get('VpcId')})
+
+    response = client.delete_vpc(**kwargs)
+    if response['ResponseMetadata']['HTTPStatusCode'] == 200:
+        demisto.results('VPC sucessfully deleted')
+
+
+def delete_internet_gateway_command(args):
+    client = aws_session(
+        region=args.get('region'),
+        roleArn=args.get('roleArn'),
+        roleSessionName=args.get('roleSessionName'),
+        roleSessionDuration=args.get('roleSessionDuration'),
+    )
+    kwargs = {}
+    if args.get('InternetGatewayId') is not None:
+        kwargs.update({'InternetGatewayId': args.get('InternetGatewayId')})
+
+    response = client.delete_internet_gateway(**kwargs)
+    if response['ResponseMetadata']['HTTPStatusCode'] == 200:
+        demisto.results('Subnet sucessfully modified')
+
+
+def describe_internet_gateway_command(args):
+    client = aws_session(
+        region=args.get('region'),
+        roleArn=args.get('roleArn'),
+        roleSessionName=args.get('roleSessionName'),
+        roleSessionDuration=args.get('roleSessionDuration'),
+    )
+    obj = vars(client._client_config)
+    kwargs = {}
+    data = []
+    output = []
+    if args.get('filters') is not None:
+        kwargs.update({'Filters': parse_filter_field(args.get('filters'))})
+    if args.get('InternetGatewayIds') is not None:
+        kwargs.update({'InternetGatewayIds': parse_resource_ids(args.get('InternetGatewayIds'))})
+
+    response = client.describe_internet_gateways(**kwargs)
+    for i, internet_gateway in enumerate(response['InternetGateways']):
+        data.append({
+            'InternetGatewayId': internet_gateway['InternetGatewayId'],
+            'OwnerId': internet_gateway['OwnerId']
+        })
+        if 'Tags' in internet_gateway:
+            for tag in internet_gateway['Tags']:
+                data[i].update({
+                    tag['Key']: tag['Value']
+                })
+        if 'Attachments' in internet_gateway:
+            for attachment in internet_gateway['Attachments']:
+                data[i].update({
+                    'State': attachment['State'],
+                    'VpcId': attachment['VpcId']
+                })
+        internet_gateway.update({'Region': obj['_user_provided_options']['region_name']})
+        output.append(internet_gateway)
+
+    try:
+        raw = json.loads(json.dumps(output, cls=DatetimeEncoder))
+    except ValueError as e:
+        return_error('Could not decode/encode the raw response - {err_msg}'.format(err_msg=e))
+    ec = {'AWS.EC2.InternetGateways(val.InternetGatewayId === obj.InternetGatewayId)': raw}
+    human_readable = tableToMarkdown('AWS EC2 Internet Gateway Ids', data)
+    return_outputs(human_readable, ec)
+
+
 """COMMAND BLOCK"""
 try:
     LOG('Command being called is {command}'.format(command=demisto.command()))