+++ Integrations/AWS-EC2/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
-Improved handling of error messages.
+ - Added the ***aws-ec2-delete-subnets***, ***aws-ec2-describe-internet-gateway***, 
+ ***aws-ec2-detach-internet-gateway***, ***aws-ec2-delete-internet-gateway***, and 
+ ***aws-ec2-delete-vpc*** commands.
 
 ## [19.9.0] - 2019-09-04
 * Added several arguments to the ***authorize_security_group_ingress*** command.